make self-reliant tests
In last versions, tests are run shuffled and the previous LDAP cases expected to be run in a specific order (with numerical names to enforce that)
Move utility functions to instance side of LDAPTest class and tests are implemented in subclasses of corresponding requests. LDAP entries are added and deleted for each test instead of relying on previous state.