move update condtion to the presenter rather than in the model
some exceptions were raised when the model is being changed between it being registered to the announcer and the time the presenter elements are initialized. this happens as the entry are loaded from the network in another process.
rather than putting the flag in the model, store the state in the presenter. even if the model is never changed (the flag never needs to be reset), the intent is clearer.